+++ source/sequence.h
@@ -4,6 +4,7 @@
 #include <functional>
 #include <algorithm>
 #include <random>
+#include <stdexcept>
 
 #include "constants.h"
 
@@ -14,7 +15,7 @@ namespace NUMCPP {
     {
     public:
 
-        Sequence(T* p0, T* p1, size_t inc) {
+        Sequence(T* p0, T* p1, int inc) {
             this->m_p0 = p0;
             this->m_p1 = p1;
             this->m_inc = inc;
@@ -28,14 +29,14 @@ namespace NUMCPP {
             this->m_size = p1 - p0;
         }
 
-        Sequence(T* p0, size_t n, size_t inc) {
+        Sequence(T* p0, int n, int inc) {
             this->m_p0 = p0;
             this->m_p1 = p0 + n * inc;
             this->m_inc = inc;
             this->m_size = n;
         }
 
-        Sequence(T* p0, size_t n) {
+        Sequence(T* p0, int n) {
             this->m_p0 = p0;
             this->m_p1 = p0 + n;
             this->m_inc = 1;
@@ -70,21 +71,21 @@ namespace NUMCPP {
             return Sequence<T>(m_p1 - m_inc, m_p0 - m_inc, -m_inc);
         }
 
-        T operator()(size_t idx)const {
+        T operator()(int idx)const {
             return *(m_p0 + idx * m_inc);
         }
 
-        T& operator()(size_t idx) {
+        T& operator()(int idx) {
             return *(m_p0 + idx * m_inc);
         }
 
         double dot(const Sequence& x)const;
 
-        size_t length()const {
+        int length()const {
             return m_size;
         }
 
-        size_t increment()const {
+        int increment()const {
             return m_inc;
         }
 
@@ -94,7 +95,7 @@ namespace NUMCPP {
 
         void rand();
 
-        void slide(size_t del);
+        void slide(int del);
 
         template<typename S>
         friend std::ostream& operator<< (std::ostream& stream, const Sequence<S>& seq);
@@ -102,14 +103,14 @@ namespace NUMCPP {
     private:
 
         T* m_p0, * m_p1;
-        size_t m_inc;
-        size_t m_size;
+        int m_inc;
+        int m_size;
 
     };
 
     template <typename T>
-    inline void Sequence<T>::slide(size_t n) {
-        size_t del = m_inc == 1 ? n : n * m_inc;
+    inline void Sequence<T>::slide(int n) {
+        int del = m_inc == 1 ? n : n * m_inc;
         m_p0 += del;
         m_p1 += del;
     }
@@ -119,12 +120,12 @@ namespace NUMCPP {
     {
     public:
 
-        DataBlock(size_t n) {
+        DataBlock(int n) {
             m_data = new T[n];
             m_size = n;
         }
 
-        DataBlock(size_t n, T* px);
+        DataBlock(int n, T* px);
 
         DataBlock(const DataBlock<T>& x);
 
@@ -138,7 +139,7 @@ namespace NUMCPP {
             return Sequence<T>(m_data, m_size);
         }
 
-        size_t length() {
+        int length() {
             return m_size;
         }
 
@@ -153,14 +154,14 @@ namespace NUMCPP {
     private:
 
         T* m_data;
-        size_t m_size;
+        int m_size;
     };
 
     template<typename T>
-    DataBlock<T>::DataBlock(size_t n, T* px) {
+    DataBlock<T>::DataBlock(int n, T* px) {
         m_data = new T[n];
         m_size = n;
-        for (size_t u = 0; u < m_size; ++u) {
+        for (int u = 0; u < m_size; ++u) {
             m_data[u] = px[u];
         }
     }
@@ -169,7 +170,7 @@ namespace NUMCPP {
     DataBlock<T>::DataBlock(const DataBlock<T>& x) {
         m_data = new T[x.m_size];
         m_size = x.m_size;
-        for (size_t u = 0; u < m_size; ++u) {
+        for (int u = 0; u < m_size; ++u) {
             m_data[u] = x.m_data[u];
         }
     }
@@ -180,7 +181,7 @@ namespace NUMCPP {
             delete[] m_data;
             m_data = new T[x.m_size];
             m_size = x.m_size;
-            for (size_t u = 0; u < m_size; ++u) {
+            for (int u = 0; u < m_size; ++u) {
                 m_data[u] = x.m_data[u];
             }
         }
@@ -223,7 +224,7 @@ namespace NUMCPP {
         if (seq.isEmpty())
             return stream;
         const T* p = seq.begin(), * p1 = seq.end();
-        size_t inc = seq.increment();
+        int inc = seq.increment();
 
         stream << *p;
         p += inc;
@@ -239,7 +240,7 @@ namespace NUMCPP {
         std::random_device rd;
         std::mt19937 mt(rd());
         std::uniform_real_distribution<double> dist(0, 1);
-        for (size_t u = 0; u < m_size; ++u)
+        for (int u = 0; u < m_size; ++u)
             m_data[u] = dist(mt);
     }
